test(unit): Add nanomsg unit tests
This adds unit tests which drive the nanomsg portion of the
library. The unit test script also now merges coverage
reports so that common portions of the library do not need
to be tested by both nano and nng tests in order to show
accurate coverage information.

+++ test/mbuf_api_static_test.c
@@ -262,7 +262,7 @@ int mbuf_api_test( ) {
 	errors += fail_not_nil( buf, "rmr_get_src returned a pointer when given a nil dest buffer" );
 
 	buf = rmr_get_src( mbuf, src_buf );
-	errors += fail_not_equal( buf, src_buf, "rmr_get_src didn't return expexted buffer pointer" );
+	errors += fail_not_equal( (int) buf, (int) src_buf, "rmr_get_src didn't return expexted buffer pointer" );
 
 
 	return errors > 0;			// overall exit code bad if errors
